tutorials/requirements.txt
notebook>=7.0.6
PyYAML>=6.0.1
